some changes

+++ src/main/java/com/chathra/fernanPharmacyBackend/repositories/SpecialitiesRepository.java
@@ -3,5 +3,5 @@ package com.chathra.fernanPharmacyBackend.repositories;
 import com.chathra.fernanPharmacyBackend.entity.Specialities;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SpecialitiesRepository extends JpaRepository<Specialities, Integer> {
+public interface SpecialitiesRepository extends JpaRepository<Specialities, Long> {
 }
